requirements.txt
fastapi==0.111.1
uvicorn==0.23.2
gtts==2.3.2
python-multipart>=0.0.7
tinydb==4.8.2
passlib==1.7.4
PyJWT==2.8.0
pydub==0.25.1
numpy==1.26.4
bcrypt==4.0.1
python-dotenv==1.0.1
email-validator==2.0.0